target/c28x: Fix loc16/loc32 hook called twice bug
A bug in the `address-mode.h` will cause the C28xLocDesc `pre_hook` and `post_hook` to be called twice in some Read-Modify-Write instructions.
In this patch, we add a flag `rmw` to the C28X_READ/WRITE_LOC16/32 macro to avoid this bug.
Hope this patch can fix the bug.